app/requirements.txt
# Spark reads Parquet via its built-in support; pyarrow/fastparquet omitted (they pull pandas).
cassandra-driver
venv-pack
pathvalidate
pyspark
tqdm
